🚚 rename: Menu 클래스 및 MenuTest 이름 변경
